feat(browser): stop emitting bundled htmx
HtmxHeadOutlet now emits only Htmxor's browser adapter. The application owns its htmx runtime and configuration, and the obsolete UseEmbeddedHtmx option is removed. Existing HtmxConfig serialization coverage is retained independently from the outlet.
